[REF] moved async controller to get the qty of valid comments per product to the right place, it was in website_variants_extra now is in website_rate_product
