+++ website_rate_product/controllers/__init__.py
@@ -1,2 +1,3 @@
 # coding: utf-8
 from . import main
+from . import async_controllers
