+++ website_rate_product/models/mail_message.py
@@ -36,6 +36,15 @@ class MailMessage(models.Model):
         res['rating'] = message.rating
         return res
 
+    @api.model
+    def _get_product_comments_qty(self, product_id):
+        mail = self.env['mail.message']
+        message_qty = mail.search_count([('res_id', '=', int(product_id)),
+                                         ('website_published', '=', True),
+                                         ('model', '=', 'product.template'),
+                                         ('type', '=', 'comment')])
+        return message_qty
+
 
 class ProductTemplate(models.Model):
     _inherit = 'product.template'
