+++ website_variants_extra/controllers/__init__.py
@@ -1,3 +1,2 @@
 # coding: utf-8
 from . import main
-from . import async_controllers
