+++ website_variants_extra/models/__init__.py
@@ -1,2 +1,2 @@
-from . import product
+# coding: utf-8
 from . import product_report
